% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchpole-package.R
\docType{package}
\name{catchpole}
\alias{catchpole}
\alias{catchpole-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/catchpole}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/catchpole/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
