% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delegates-pal.R
\docType{data}
\name{delegates_pal}
\alias{delegates_pal}
\title{WSJ palette for the candidates likely getting delegates}
\format{An object of class \code{character} of length 9.}
\usage{
delegates_pal
}
\description{
WSJ palette for the candidates likely getting delegates
}
\keyword{datasets}
