% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-candidates.R
\name{expand_candidates}
\alias{expand_candidates}
\title{Expand the state totals per delegate into an indexed data frame}
\usage{
expand_candidates(delegates = read_delegates())
}
\arguments{
\item{delegates}{a data frame with \code{state}, \code{candidate}, and
\code{delegates} count or use the default which gets the
current data as long as the WSJ keeps it alive.}
}
\description{
Expand the state totals per delegate into an indexed data frame
}
