% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-catchpole.R
\name{gg_catchpole}
\alias{gg_catchpole}
\title{Plot a U.S. Democratic Delegates Cargogram (2020)}
\usage{
gg_catchpole(delegates = read_delegates())
}
\arguments{
\item{delegates}{a data frame with \code{state}, \code{candidate}, and
\code{delegates} count or use the default which gets the
current data as long as the WSJ keeps it alive.}
}
\description{
Plot a U.S. Democratic Delegates Cargogram (2020)
}
