% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-delegates.R
\name{read_delegates}
\alias{read_delegates}
\title{Retrieves current U.S. State delegate assignments for 2020 from the WSJ}
\usage{
read_delegates()
}
\description{
Retrieves current U.S. State delegate assignments for 2020 from the WSJ
}
\references{
\url{https://asset.wsj.net/wsjnewsgraphics/election/2020/delegates.json}
}
