% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchpole-package.R
\docType{package}
\name{catchpole}
\alias{catchpole}
\alias{catchpole-package}
\title{2020 Democrat POTUS Delegate Map Tracker}
\description{
Tools to build a 2020 (D) delegate cartogram
}
\section{Credit}{


The hard work was done by WSJ's Brian McGill (\href{https://twitter.com/brian_mcgill/}{@brian_mcgill}) who \strong{manually crafted} the base cartogram over \strong{days} then translated that into the beautiful and illuminating creation that is the WSJ online delegate tracker: \url{https://www.wsj.com/graphics/elections/2020/track-the-delegate-count/}.

Brians's tweetnouncement: \url{https://twitter.com/brian_mcgill/status/1222556373864648704?ref_src=twsrc\%5Etfw}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/catchpole}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/catchpole/issues}
}

}
\author{
Bob Rudis (bob@rud.is)

Brian McGill
}
\keyword{internal}
